+++ contexts/types.ts
@@ -1,0 +1,12 @@
+export interface IProviderContext {
+  cords
+  city
+  tab
+  platform
+  message
+  setCords
+  setCity
+  setTab
+  setPlatform
+  setMessage
+}
