+++ contexts/utils.tsx
@@ -1,0 +1,7 @@
+import { Context, useContext } from 'react'
+
+export function createUseContextHook<T>(ctx: Context<T>) {
+  return function useSpecificContext(): T {
+    return useContext(ctx)
+  }
+}
