+++ pages/_app.tsx
@@ -1,11 +1,14 @@
 import '../styles/globals.css'
 import type { AppProps } from 'next/app'
 import DocumentHead from '../components/DocumentHead'
+import DataProvider from '../contexts'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <DocumentHead pageTitle={pageProps.title}>
-      <Component {...pageProps} />{' '}
+      <DataProvider>
+        <Component {...pageProps} />
+      </DataProvider>
     </DocumentHead>
   )
 }
